Remove unused modules and classes from wrr_utils
- Deleted DNJ optimizer implementation from dnj.py
- Removed hexagonal grid functions from hexagonal_grid.py
- Eliminated polygon manipulation functions from polygons.py
- Cleared results processing and optimization logic from results.py
